+++ packages/data-types/src/lib/interfaces/animal-shop-data.interface.ts
@@ -15,4 +15,5 @@ export interface AnimalShopData {
     readableRequirement: TranslationKey | null,
     readableName: TranslationKey | null,
     requirements?: RequirementEntry
+    variant?: string
 }