+++ packages/guide/src/app/locations/components/ranch/ranch.component.ts
@@ -42,9 +42,10 @@ export class RanchComponent extends BaseShopComponent {
         if (!values) return null;
         const [animals, shopData] = values;
 
-        return shopData.map(stopEntry => {
-            const {animalKey, ...rest} = stopEntry;
-            return {...rest, animal: animals.find(animal => animal.key === animalKey)}
+        return shopData.map(shopEntry => {
+            const {animalKey, ...rest} = shopEntry;
+            const animal = animals.find(animal => animal.key === animalKey);
+            return {...rest, animal, iconName: animal?.variants.find(variant => variant.variant === shopEntry.variant)?.icons?.adult ?? animal?.variants?.[0]?.icons?.adult ?? null}
         });
     })
     protected selectedAnimal?: MappedAnimalShopData