+++ packages/guide/src/app/locations/types/mapped-animal-shop-data.type.ts
@@ -1,3 +1,3 @@
 import { AnimalData, AnimalShopData } from "@ci/data-types";
 
-export type MappedAnimalShopData = Omit<AnimalShopData, 'animalKey'> & { animal: AnimalData | undefined }
+export type MappedAnimalShopData = Omit<AnimalShopData, 'animalKey'> & { animal: AnimalData | undefined, iconName: string | null }