+++ packages/pak-files-parser/src/interfaces/raw-data-interfaces/raw-animal-shop-data.interface.ts
@@ -14,5 +14,7 @@ export interface RawAnimalShopData {
     "readableCategory": SourceString,
     "readableRequirement": SourceString,
     "readableName": SourceString,
-    "priority": number
+    "priority": number;
+    "isStarletAnimalVariant": boolean,
+    "starletAnimalVariant": string
 }